+++ src/ploomber/clients/storage/aws.py
@@ -155,6 +155,10 @@ class S3Client(AbstractStorageClient):
     ploomber.clients.GCloudStorageClient :
         Client for uploading products to Amazon Google Cloud Storage
 
+    Notes
+    -----
+    If a notebook (or script) task fails, the partially executed ``.ipynb``
+    file will be uploaded using this client.
     """
     @requires(['boto3', 'botocore'], name='S3Client')
     def __init__(self,
@@ -265,3 +269,7 @@ class S3Client(AbstractStorageClient):
         return (f'{type(self).__name__}(bucket_name={self._bucket_name!r}, '
                 f'parent={self._parent!r}, '
                 f'path_to_project_root={str(self._path_to_project_root)!r})')
+
+    @property
+    def parent(self):
+        return str(self._parent)