+++ src/ploomber/static_analysis/jinja.py
@@ -40,9 +40,13 @@ class JinjaExtractor:
         """
         attr = self.ast.find_all(Getattr)
         item = self.ast.find_all(Getitem)
+
         return set([
             obj.arg.as_const() if isinstance(obj, Getitem) else obj.attr
-            for obj in chain(attr, item) if obj.node.name == variable
+            # iterate over get attribute and get item
+            for obj in chain(attr, item)
+            # only check variable access
+            if hasattr(obj.node, 'name') and obj.node.name == variable
         ]) or None
 
     def find_variable_assignment(self, variable):