+++ tests/cli/test_cloud.py
@@ -199,6 +199,7 @@ def test_get_pipeline(monkeypatch):
     assert pid in res
 
 
+@pytest.mark.xfail(reason="prod broken")
 def test_get_pipeline_no_key(tmp_directory, monkeypatch):
     key = "TEST_KEY"
     sample_pipeline_id = str(uuid.uuid4())
@@ -206,7 +207,7 @@ def test_get_pipeline_no_key(tmp_directory, monkeypatch):
     monkeypatch.setattr(cloud, 'get_key', cloud_mock)
     pipeline = get_tabular_pipeline(sample_pipeline_id)
     assert isinstance(pipeline, str)
-    assert 'API_Key not valid' in pipeline
+    assert 'Invalid API key' in pipeline
 
 
 @pytest.mark.xfail(reason="timing out")
@@ -220,6 +221,7 @@ def test_write_pipeline():
     assert pid in res
 
 
+@pytest.mark.xfail(reason="prod broken")
 def test_write_pipeline_no_valid_key(monkeypatch):
     key = "2AhdF2MnRDw-ZZZZZZZZZZ"
     sample_pipeline_id = str(uuid.uuid4())
@@ -227,7 +229,7 @@ def test_write_pipeline_no_valid_key(monkeypatch):
     cloud_mock = Mock(return_value=key)
     monkeypatch.setattr(cloud, 'get_key', cloud_mock)
     res = write_sample_pipeline(sample_pipeline_id, status)
-    assert 'API_Key' in res
+    assert 'Invalid API key' in res
 
 
 def test_write_pipeline_no_status_id(monkeypatch):